+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- errors: add `TrySendError::map` and `RequestError::map`.
 
 ## [0.1.32] - 2021-12-21
 ### Added